refactor: correct reducer functions
